% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_FR.R
\name{calculate_FR}
\alias{calculate_FR}
\title{Calculate fuel ratio}
\usage{
calculate_FR(VM = NA, FC = NA)
}
\arguments{
\item{VM}{Volatile matter}

\item{FC}{Fixed carbon}
}
\description{
This function calculates coal feul ratio which is required for coal boiler design
}
\details{
Forumla for determining fuel ratio. The forumla is:

\deqn{FR = FC/VM}

Where FC and VM are at the same basis
}
\examples{
calculate_FR(FC=65,VM=18)
}
\references{
TBA
}
