% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_Mis.R
\name{calculate_Mis}
\alias{calculate_Mis}
\title{Calculate in situ moisture (Mis)}
\usage{
calculate_Mis(model, Ashd = NA, VMdaf = NA, Cdaf = NA, Mad = NA,
  Hdaf = NA, MHC = NA, EM = NA, Codmmf = NA, Lipmmf = NA)
}
\arguments{
\item{model}{The character string indicating the in situ moisture calculation formula to use. Must be one of "M1A", "M1B", "M2A".}

\item{Ashd}{Ash \% (dry basis) (required for all models).}

\item{VMdaf}{Volatile Matter \% (dry ash free basis).}

\item{Cdaf}{Carbon from Ultimates \% (dry ash free basis).}

\item{Mad}{Moisture \% (air dried)}

\item{Hdaf}{Hydrogen \% (dry ash free)}

\item{MHC}{Moisture Holding Capacity \% for higher rank coals (AS1038.17-2000)}

\item{EM}{Equilibrium Moisture \%}

\item{Codmmf}{Carbon, organic (dry, mineral matter-free)}

\item{Lipmmf}{Liptinite (mineral matter-free)}
}
\description{
This function directly calculates in situ moisture with option to use one of three alternate models/formula.
}
\details{
Valid model text strings include any of:

'M1A', 'M1B', 'M2A'

Model "M1A" is taken from ACARP report C0041 (eq 5.1). It requires Equilibrium Moisture as the only argument and is derived by the formula:

\deqn{1.117*EM+0.317}

Model "M1B" is taken from ACARP report C0041 (eq 5.9). It requires Moisture Holding Capacity, Organic Carbon (dmmf) and Liptinite (mmf) as arguments and is derived by the forumla:

\deqn{1.2283*MHC+12.5334*Codmmf*-0.0742*Codmmf^2-0.9855*Lipmmf+0.0426*Lipmmf^2-526.56}
}
\examples{
calculate_Mis("M1A",EM=5)

}
\references{
"Blah"
}
