% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_phos.R
\name{calculate_phos}
\alias{calculate_phos}
\title{Calculate phosphorus in coal}
\usage{
calculate_phos(A = NA, P2O5 = NA)
}
\arguments{
\item{A}{Ash \%}

\item{P2O5}{P2O5 \% from ash analysis}
}
\description{
This function calculates phosphorus in coal from ash analysis
}
\details{
Uses the below formula to calculate phosphorus in coal.

\deqn{0.43642*A*P2O5/ 100}

Where A and P2O5 are at the same basis
}
\examples{
calculate_phos(A=15,P2O5=5)
}
