% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bases.R
\name{convert_bases}
\alias{convert_bases}
\title{Convert (coal) bases}
\usage{
convert_bases(x, fromto, im = NA, tm = NA, nm = NA, Ashd = NA)
}
\arguments{
\item{x}{The analyses to be converted, for example ash, volatile matter, fixed carbon ect. \emph{Character}, \emph{required}}

\item{fromto}{The text string indicating the desired conversion, \emph{Character}, \emph{required}}

\item{im}{Inherent moisture \% (commonly refered to as air dried moisture). \emph{Numeric}}

\item{tm}{Total moisture \% (the free and air dried moisture). \emph{Numeric}}

\item{nm}{Nomiated moisture. \emph{Numeric}}

\item{Ashd}{Ash \% (dry basis).  \emph{Numeric}}
}
\description{
This function enables the conversion of coal analyses to a desired moisture basis
}
\details{
The function allows for the conversion of coal analyses from or to (fromto) desired mositure basis including:
\itemize{
\item From air dried basis to dry ("adtodry") or from dry to air dried ("drytoad")
\item From air dried basis to as received ("adtoar") or from as received to air dried ("artoad")
\item From air dried basis to any nominated (fixed) basis ("adtonm")
\item From as received to dry ("artodry") or from dry to as received ("drytoar")
\item 
}

Valid fromto text strings include any of:

"adtodry","adtoar", "artoad","adtonm", "artodry","drytoad","drytoar","drytodaf"

\emph{"adtodry"}
}
\examples{
CoalCalcs::convert_bases(13.5,"adtonm",4,NA,5,NA)
}
