% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_CV_units.R
\name{convert_CV_units}
\alias{convert_CV_units}
\title{Convert Calorific Value (CV) units}
\usage{
convert_CV_units(fromto, CV)
}
\arguments{
\item{fromto}{Text string indicating desired conversion. May be one of "kcaltomj", "kcaltobtu", "mjtokcal", "mjtobtu", "btutomj", "btutokcal"}

\item{CV}{Calorific Value}
}
\description{
This function converts Calorific Value between units kcal/kg, Mj/kg, Btu/lb and Mj/kg to J/g
}
\details{
Uses conversion factors as outlined in AS 1038.16 to convert Calorific value units which may be reproted in kcal/kg, Mj/kg, or Btu/lb

\deqn{RDis = (RDad*(100-Mad))/(100+RDad*(Mis-Mad)-Mis)}
}
\examples{
#Convert CV units in Mj/kg to kcal/kg and Btu/lb
l = c("mjtokcal", "mjtobtu")
sapply(l, function (x) CoalCalcs::convert_CV_units(x,40))
}
\references{
AS 1038.16 Australian Standard. Coal and Coke - Analysis and testing - Higher Rank Coals - Reporting of Test Results

MineSkill Australia, 2001, Understanding Coal Quality, Willson, W., Daneille, F., Gibson, L. (eds).
}
