% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_RDadtoRDdry.R
\name{convert_RDadtoRDdry}
\alias{convert_RDadtoRDdry}
\title{Convert relative density air dried to dry basis}
\usage{
convert_RDadtoRDdry(RDad = NA, Mad = NA)
}
\arguments{
\item{RDad}{Relative density (air dried)}

\item{Mad}{Air dried moisture \%}
}
\description{
This function converts relative density (air dried) (RDad) to dry density (RDdry) using Prestons Sanders equation
}
\details{
Uses Preston Sanders formula (Preston and Sanders (1993)) to convert relative density (air dried) (RDad) to insitu density (RDdry). The forumla is:

\deqn{RDdry = RDad*(100-Mad)/(100-RDad*Mad)}

The formula can alternatively be written:

\deqn{RDdry =RDad*(1-0.01*Mad))/(1-(RDad*0.01*Mad))}
}
\examples{
convert_RDadtoRDdry(RDad=1.4,Mad=4.5)
}
\references{
Preston, K.B., Sanders, R.H., 1993. Estimating the in-situ relative density of coal. Aust. Coal Geol. 9.
}
