% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_RDadtoRDis.R
\name{convert_RDadtoRDis}
\alias{convert_RDadtoRDis}
\title{Convert relative density (air dried) (RDad) to insitu density (RDis)}
\usage{
convert_RDadtoRDis(RDad = NA, Mad = NA, Mis = NA)
}
\arguments{
\item{RDad}{Relative density (air dried) (required)}

\item{Mad}{Air dried moisture \%}

\item{Mis}{Insitu moisture}
}
\description{
This function converts relative density (air dried) (RDad) to insitu density (RDis) using Prestons Sanders equation
}
\details{
Uses Preston Sanders formula (Preston and Sanders (1993)) to convert relative density (air dried) (RDad) to insitu density (RDis). The forumla is:

\deqn{RDis = (RDad*(100-Mad))/(100+RDad*(Mis-Mad)-Mis)}
}
\examples{
convert_RDadtoRDis(RDad=1.4,Mad=3.4,Mis=5.2)
}
\references{
Preston, K.B., Sanders, R.H., 1993. Estimating the in-situ relative density of coal. Aust. Coal Geol. 9.
}
