% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bases.R
\name{convert_bases}
\alias{convert_bases}
\title{Convert (coal) bases}
\usage{
convert_bases(
  x,
  fromto,
  im = NA,
  tm = NA,
  nm = NA,
  Ashd = NA,
  Ashad = NA,
  Ashar = NA,
  Sdry = NA
)
}
\arguments{
\item{x}{The analyses to be converted, for example ash, volatile matter, fixed carbon ect.}

\item{fromto}{The text string indicating the desired conversion}

\item{im}{Inherent moisture \% (commonly referred to as air dried moisture)}

\item{tm}{Total moisture \% (the free and air dried moisture)}

\item{nm}{Nominated moisture}

\item{Ashd}{Ash \% (dry basis)}

\item{Ashad}{Ash \% (air dry basis)}

\item{Ashar}{Ash \% (as received basis)}

\item{Sdry}{Total Sulfur \% (dry basis)}
}
\description{
This function enables the conversion of coal analyses to a desired moisture basis
}
\details{
The function allows for the conversion of common coal analyses from or to desired moisture basis. This function is NOT suitable for relative density basis conversion. Bases conversions include:
\itemize{
\item From air dried basis to dry ("adtodry") or from dry to air dried ("drytoad")
\item From air dried basis to as received ("adtoar") or from as received to air dried ("artoad")
\item From air dried basis to any nominated (fixed) basis ("adtonm")
\item From as received to dry ("artodry") or from dry to as received ("drytoar")
\item From dry to dry mineral matter free (drytodmmf) using Parr formula
}

Valid fromto text strings include any of:

"adtodry","adtoar", "artoad","adtonm", "artodry","drytoad","drytoar","drytodaf"
}
\examples{
df = data.frame(M=6.8,A=12.3, VM=36.2,FC=44.2, S=0.5)
df$Adry =CoalCalcs::convert_bases(df$A,"adtodry", im=df$M)
df$VMdry=CoalCalcs::convert_bases(df$VM,"adtodry", im=df$M)
df$Adry =CoalCalcs::convert_bases(df$A,"adtodry", im=df$M)
df$FCdry= CoalCalcs::convert_bases(df$FC, "adtodry", im=df$M)
df$Sdry=CoalCalcs::convert_bases(df$S,"adtodry", im=df$M)
df$VMdaf = CoalCalcs::convert_bases(df$VMdry,"drytodaf", im=df$M,Ashd = df$Adry)
df$FCdaf = CoalCalcs::convert_bases(df$FCdry,"drytodaf", im=df$M,Ashd = df$Adry)
df$FCdmmf = CoalCalcs::convert_bases(df$FCdry,"drytodmmf", Sdry=df$Sdry, Ashd = df$Adry)

}
