% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_RDadtoRDis.R
\name{convert_RDadtoRDis}
\alias{convert_RDadtoRDis}
\title{Convert relative density (air dried) (RDad) to insitu density (RDis)}
\usage{
convert_RDadtoRDis(RDad = NA, im = NA, is = NA)
}
\arguments{
\item{RDad}{Relative density (air dried) (required)}

\item{im}{Inherent moisture \% (commonly refered to as air dried moisture)}

\item{is}{Insitu moisture \% (required for "prestons" method)}
}
\description{
This function converts relative density (air dried) (RDad) to insitu density (RDis) using Prestons Sanders equation
}
\details{
Uses Preston Sanders formula to convert relative density (air dried) (RDad) to insitu density (RDis). The forumla is:

 RDis = (RDad*(100-im))/(100+RDad*(is-im)-is)
}
\examples{
convert_RDadtoRDis(1.4,3.4,5.2)
}
